samples/python/agents/search_agnt/pyproject.toml
[project]
name = "a2a-sample-agent-searchengine"
version = "0.1.0"
description = "Sample Search Agent with A2A Protocol"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "a2a-samples",
    "click>=8.1.8",
    "httpx>=0.28.1",
    "langchain>=0.1.17",
    "langchain-openai>=0.1.5",
    "langchain-community>=0.0.27",
    "tavily-python>=0.3.2",
    "langgraph>=0.3.18",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv.sources]
a2a-samples = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"